tests voice init
